requirements.txt
aniso8601==9.0.1
argon2-cffi==21.1.0
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.0.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.23.1
pycparser==2.21
PyJWT==2.3.0
pymongo==3.12.1
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
WTForms==3.0.0
